+++ R/plotHeatmap.R
@@ -59,9 +59,9 @@
 #' @importFrom stats "reorder"
 #' @importFrom pheatmap "pheatmap"
 #' @importFrom stats "na.omit"
-#' @importFrom viridis "viridis"
+#' @importFrom grDevices "hsv"
 plotHeatmap <- function (data, pseudocount = 1e-6, 
-                         colours = viridis::viridis(100), 
+                         colours = grDevices::hsv(1,1,seq(0,1,length.out = 12)),
                          cluster = TRUE, scale = "none", 
                          clustering_method = "complete", 
                          row.names = TRUE)